Fix nrepl-load-file-form using nrepl-format-vector.
Not terribly happy with this, but it's better than embedding big
strings of Clojure code. The primary problem is that while you can
"unquote" in vectors with %s, it only works at the top level of the
vector, which makes for the awkwardness of this:

(let [n %s ns (symbol n)] ...)

Also we work around the fact that dotted symbols are escaped by not
using dotted symbol literals but calling Clojure's `symbol' function
instead. Not thrilled, but there you have it.

Perhaps there's a better solution, but this is all I have for now.